[test.function_expression] fixed some warning

+++ dune/stuff/test/function_expression.cc
@@ -15,10 +15,10 @@ struct RunExpressionBaseTest
   {
     using namespace Dune;
     using namespace Dune::Stuff;
-    typedef double    DomainFieldType;
-    static const int  dimDomain = DimDomain::value;
-    typedef double    RangeFieldType;
-    static const int  dimRange = DimRange::value;
+    typedef double      DomainFieldType;
+    static const size_t dimDomain = DimDomain::value;
+    typedef double      RangeFieldType;
+    static const size_t dimRange = DimRange::value;
 
     typedef FunctionExpressionBase< DomainFieldType, dimDomain, RangeFieldType, dimRange > FunctionType;
     const std::string variable = "x";
